+++ packages/reactivity-core/src/types.ts
@@ -99,6 +99,29 @@ export interface ExternalReactive<T> extends ReadonlyReactive<T> {
 }
 
 /**
+ * A function that shall return `true` if `a` and `b` are considered equal, `false` otherwise.
+ *
+ * @group Primitives
+ */
+export type EqualsFunc<T> = (a: T, b: T) => boolean;
+
+/**
+ * Options that can be passed when creating a new signal.
+ *
+ * @group Primitives
+ */
+export interface ReactiveOptions<T> {
+    /**
+     * Shall return `true` if the two values are considered equal.
+     *
+     * Reactive assignments using a new value equal to the current value
+     * will be ignored.
+     * By default, `Object.is` is used to compare values.
+     */
+    equal?: EqualsFunc<T>;
+}
+
+/**
  * A handle returned by various functions to dispose of a resource,
  * such as a watcher or an effect.
  *
@@ -221,7 +244,7 @@ export interface WatchOptions<T> {
      *
      * By default, an implementation based on object identity is used.
      */
-    equal?(prev: T, next: T): boolean;
+    equal?: EqualsFunc<T>;
 }
 
 /**